first version that passes all existing tests
